fixed update and delete methods

+++ src/main/java/mate/academy/intro/service/BookServiceImpl.java
@@ -1,5 +1,6 @@
 package mate.academy.intro.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import mate.academy.intro.dto.BookDto;
@@ -38,7 +39,11 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteById(Long id) {
-        bookRepository.deleteById(id);
+        if (bookRepository.existsById(id)) {
+            bookRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Can not find book with id: " + id);
+        }
     }
 
     @Override
@@ -46,6 +51,10 @@ public class BookServiceImpl implements BookService {
         Book book = bookRepository.findById(id)
                 .orElseThrow(()
                         -> new EntityNotFoundException("Cant find book by id " + id));
+        book.setTitle(bookDto.getTitle());
+        book.setAuthor(bookDto.getAuthor());
+        book.setIsbn(bookDto.getIsbn());
+        book.setPrice(BigDecimal.valueOf(bookDto.getPrice()));
         return bookMapper.bookToBookDto(bookRepository.save(book));
     }
 }
